refactor: move configuration helpers to new module
Moved from Application to Configuration for WatcherRPC.
