+++ apps/omg_watcher_rpc/lib/configuration.ex
@@ -12,23 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-defmodule OMG.WatcherRPC.Web.Response do
+defmodule OMG.WatcherRPC.Configuration do
   @moduledoc """
-  Prepares the response into the expected result/data format.
-
-  Contains only the behaviours specific to the watcher.
-  For the generic response, see `OMG.Utils.HttpRPC.Response`.
+  Provides access to applications configuration
   """
+  @app :omg_watcher_rpc
 
-  alias OMG.WatcherRPC.Configuration
+  @spec version() :: String.t()
+  def version() do
+    OMG.Utils.HttpRPC.Response.version(@app)
+  end
 
-  @doc """
-  Adds "version" and "service_name" to the response map.
-  """
-  @spec add_app_infos(map()) :: %{version: String.t(), service_name: String.t()}
-  def add_app_infos(response) do
-    response
-    |> Map.put(:version, Configuration.version())
-    |> Map.put(:service_name, Configuration.service_name())
+  @spec service_name() :: atom()
+  def service_name() do
+    Application.get_env(@app, :api_mode)
   end
 end
